Tested changing color by slowly increasing RGB values

+++ demo_color_mixer/src/main/java/ee5/demo_color_mixer/Calculator.java
@@ -10,11 +10,11 @@ public class Calculator  {
 
     }
 
-    public double calcDistance() {
-        double distance;
+    public int calcDistance() {
+        int distance;
         double xDiff = x2 - x1;
         double yDiff = y2 - y1;
-        distance = Math.sqrt(xDiff*xDiff + yDiff*yDiff);
+        distance = (int) Math.sqrt(xDiff*xDiff + yDiff*yDiff);
         return distance;
     }
